refactor: move styled-components files to a separate folder within styles
